Two Sample Kolmogorov-Smirnov Test

Performs two sample Kolmogorov-Smirnov test on data vectors.

Super classes

LearnNonparam::Permutation Test -> LearnNonparam::Two Sample Permutation Test -> Two Sample Kolmogorov-Smirnov Test

Methods

Public methods

KolmogorovSmirnov$new()

KolmogorovSmirnov$plot_ecdfs()

Inherited methods

LearnNonparam::Permutation Test$feed()

LearnNonparam::Permutation Test$plot_hist()

Method new()

Create a new KolmogorovSmirnov object.

Usage

KolmogorovSmirnov$new(n_permu = NULL)

Arguments

n_permu

an integer specifying how many permutations should be used to construct the permutation distribution. If NULL (default) then all permutations are used.

Returns

A KolmogorovSmirnov object.

Method plot_ecdfs()

Draw the empirical cumulative distribution function of each of the data fed.

Usage

KolmogorovSmirnov$plot_ecdfs()

Returns

The object itself (invisibly).