Multiple Comparison Based on T Test

Performs t test based multiple comparison on data vectors.

Super classes

LearnNonparam::Permutation Test -> LearnNonparam::K Sample Permutation Test -> LearnNonparam::Multiple Comparison -> Multiple Comparison (t test)

Active bindings

bonferroni

Whether to apply bonferroni adjustment.

Methods

Public methods

MultiCompT$new()

Inherited methods

LearnNonparam::Permutation Test$feed()

LearnNonparam::Permutation Test$plot_hist()

Method new()

Create a new MultiCompT object.

Usage

MultiCompT$new(
  type = c("permu", "approx"),
  bonferroni = TRUE,
  signif_level = 0.05,
  n_permu = NULL,
  scoring = c("none", "rank", "vw", "expon")
)

Arguments

type

a character string specifying the way to calculate p-values, must be one of "permu" (default) or "approx".

bonferroni

a logical indicating whether to apply bonferroni adjustment.

signif_level

a numeric value between zero and one giving the significance level.

n_permu

an integer specifying how many permutations should be used to construct the permutation distribution. If NULL (default) then all permutations are used.

scoring

a character string specifying which scoring system to be used, must be one of "none" (default), "rank, "vw" or "expon".

Returns

A MultiCompT object.