Inference on Cumulative Distribution Function

Performs statistical inference on population cumulative distribution function.

Super classes

LearnNonparam::PermuTest -> LearnNonparam::OneSampleTest -> CDF

Methods

Public methods

CDF$new()

CDF$plot()

Inherited methods

LearnNonparam::PermuTest$print()

LearnNonparam::PermuTest$test()

Method new()

Create a new CDF object.

Usage

CDF$new(conf_level = 0.95)

Arguments

conf_level

a number specifying confidence level of the confidence bounds.

Returns

A CDF object.

Method plot()

Plot the estimate and confidence bounds for population cumulative distribution function.

Usage

CDF$plot(style = c("graphics", "ggplot2"))

Arguments

style

a character string specifying which package to use.

Returns

The object itself (invisibly).