Two Sample Permutation Test Based on Correlation Coefficients

Performs correlation coefficient based two sample permutation test on data vectors.

Super classes

LearnNonparam::Permutation Test -> LearnNonparam::Two Sample Permutation Test -> LearnNonparam::Paired Two Sample Permutation Test -> LearnNonparam::Two Sample Permutation Test for Association -> Two Sample Permutation Test (correlation coefficient)

Methods

Public methods

Correlation$new()

Inherited methods

LearnNonparam::Permutation Test$feed()

LearnNonparam::Permutation Test$plot_hist()

Method new()

Create a new Correlation object.

Usage

Correlation$new(
  type = c("permu", "approx"),
  method = c("pearson", "kendall", "spearman"),
  alternative = c("two_sided", "less", "greater"),
  n_permu = NULL
)

Arguments

type

a character string specifying the way to calculate p-values, must be one of "permu" (default) or "approx".

method

a character string indicating which correlation coefficient is to be computed. Must be one of "pearson" (default), "kendall", or "spearman".

alternative

a character string specifying the alternative hypothesis, must be one of "two_sided" (default), "greater" or "less".

n_permu

an integer specifying how many permutations should be used to construct the permutation distribution. If NULL (default) then all permutations are used.

Returns

A Correlation object.