One-Way Test for Equal Means

Performs F statistic based one-way test on samples.

Super classes

LearnNonparam::PermuTest -> LearnNonparam::KSampleTest -> OneWay

Methods

Public methods

OneWay$new()

Inherited methods

LearnNonparam::PermuTest$plot()

LearnNonparam::PermuTest$print()

LearnNonparam::PermuTest$test()

Method new()

Create a new OneWay object.

Usage

OneWay$new(type = c("permu", "asymp"), n_permu = 10000)

Arguments

type

a character string specifying the way to calculate the p-value.

n_permu

an integer indicating number of permutations for the permutation distribution. If set to 0, all permutations will be used.

Returns

A OneWay object.

Examples

# \donttest{
t <- pmt(
    "ksample.oneway", n_permu = 0
)$test(Table3.1.2)$print()
#> 
#>  	 One-Way Test for Equal Means 
#> 
#> scoring: none    type: permu(756756)    method: default
#> statistic = 11552.89, p-value = 0.05127412

t$type <- "asymp"
t
#> 
#>  	 One-Way Test for Equal Means 
#> 
#> scoring: none    type: asymp    method: default
#> statistic = 3.781445, p-value = 0.05327082
# }