PermuTest Class

This is the abstract base class for permutation test objects. Note that it is not recommended to create objects of this class directly.

Active bindings

type

The type of the test.

method

The method used.

scoring

The scoring system used.

null_value

The value of the parameter in the null hypothesis.

alternative

The alternative hypothesis.

conf_level

The confidence level of the interval.

n_permu

The number of permutations used.

data

The data.

statistic

The test statistic.

p_value

The p-value.

estimate

The estimated parameter.

ci

The confidence interval.

Methods

Public methods

PermuTest$new()

PermuTest$test()

PermuTest$print()

PermuTest$plot()

Method new()

Create a new PermuTest object. Note that it is not recommended to create objects of this class directly.

Usage

PermuTest$new(
  null_value = 0,
  alternative = c("two_sided", "less", "greater"),
  n_permu = 0L,
  conf_level = 0.95,
  scoring = c("none", "rank", "vw", "expon")
)

Arguments

null_value

a number specifying the value of the parameter in the null hypothesis.

alternative

a character string specifying the alternative hypothesis.

n_permu

an integer specifying how many permutations should be used to construct the permutation distribution. If set to zero (default) then all permutations are used.

conf_level

a number specifying confidence level of the interval.

scoring

a character string specifying which scoring system to be used.

Returns

A PermuTest object.

Method test()

Perform test on data.

Usage

PermuTest$test(...)

Arguments

...

data.

Returns

The object itself (invisibly).

Method print()

Print the results of the test.

Usage

PermuTest$print(digits = getOption("digits"))

Arguments

digits

an integer specifying the minimum number of significant digits to be printed in values.

Returns

The object itself (invisibly).

Method plot()

Plot histogram(s) of the permutation distribution. Note that it works only if the test's type is "permu".

Usage

PermuTest$plot(style = c("graphics", "ggplot2"), ...)

Arguments

style

a character string specifying which package to be used.

...

extra parameters passed to graphics::hist or ggplot2::stat_bin.

Returns

The object itself (invisibly).