Quantile Test

Performs quantile test on a single sample. In addition, an estimation and a confidence interval for the desired quantile will be calculated.

Super classes

LearnNonparam::PermuTest -> LearnNonparam::OneSampleTest -> Quantile

Active bindings

prob

The probability associated with the quantile.

correct

Whether to apply continuity correction.

Methods

Public methods

Quantile$new()

Inherited methods

LearnNonparam::PermuTest$print()

LearnNonparam::PermuTest$test()

LearnNonparam::OneSampleTest$plot()

Method new()

Create a new Quantile object.

Usage

Quantile$new(
  type = c("asymp", "exact"),
  alternative = c("two_sided", "less", "greater"),
  null_value = 0,
  conf_level = 0.95,
  prob = 0.5,
  correct = TRUE
)

Arguments

type

a character string specifying the way to calculate the p-value.

alternative

a character string specifying the alternative hypothesis.

null_value

a number indicating the hypothesized value of the quantile.

conf_level

a number between zero and one indicating the confidence level to use.

prob

a number between zero and one indicating the probability associated with the quantile.

correct

a logical indicating whether to apply continuity correction in the normal approximation for the p-value.

Returns

A Quantile object.