Siegel-Tukey Test

Performs Siegel-Tukey test on samples.

Super classes

LearnNonparam::PermuTest -> LearnNonparam::TwoSampleTest -> LearnNonparam::TwoSampleLocationTest -> LearnNonparam::Wilcoxon -> SiegelTukey

Methods

Public methods

SiegelTukey$new()

Inherited methods

LearnNonparam::PermuTest$plot()

LearnNonparam::PermuTest$print()

LearnNonparam::PermuTest$test()

Method new()

Create a new SiegelTukey object.

Usage

SiegelTukey$new(
  type = c("permu", "asymp", "exact"),
  alternative = c("two_sided", "less", "greater"),
  n_permu = 10000,
  correct = TRUE
)

Arguments

type

a character string specifying the way to calculate the p-value.

alternative

a character string specifying the alternative hypothesis.

n_permu

an integer indicating number of permutations for the permutation distribution. If set to zero (default) then all permutations are used.

correct

a logical indicating whether to apply continuity correction in the normal approximation for the p-value.

Returns

A SiegelTukey object.