Multiple Comparison Based on Studentized Statistic

Performs studentized statistic based multiple comparison on samples.

Super classes

LearnNonparam::PermuTest -> LearnNonparam::KSampleTest -> LearnNonparam::MultipleComparison -> Studentized

Methods

Public methods

Studentized$new()

Inherited methods

LearnNonparam::PermuTest$plot()

LearnNonparam::PermuTest$print()

LearnNonparam::PermuTest$test()

Method new()

Create a new Studentized object.

Usage

Studentized$new(
  type = c("permu", "asymp"),
  method = c("bonferroni", "tukey"),
  scoring = c("none", "rank", "vw", "expon"),
  conf_level = 0.95,
  n_permu = 10000
)

Arguments

type

a character string specifying the way to calculate the p-value.

method

a character string specifying whether to use Bonferroni's method or Tukey's HSD method.

scoring

a character string specifying the scoring system.

conf_level

a number between zero and one indicating the family-wise confidence level to use.

n_permu

an integer indicating number of permutations for the permutation distribution. If set to 0, all permutations will be used.

Returns

A Studentized object.

Examples

t <- pmt(
    "multcomp.studentized", method = "bonferroni"
)$test(Table3.3.1)$print()
#> 
#> 	 Multiple Comparison Based on Studentized Statistic 
#> 
#> scoring: none    type: permu(10000)    method: bonferroni
#> 
#> family-wise confidence level: 95%
#> 
#>                           statistic p-value  
#> location_1 ~ location_2  0.02737145  0.9754  
#> location_1 ~ location_3 -1.77914406  0.0844  
#> location_1 ~ location_4 -3.41048230  0.0018 *
#> location_2 ~ location_3 -1.80651550  0.0868  
#> location_2 ~ location_4 -3.43785375  0.0018 *
#> location_3 ~ location_4 -1.63133824  0.1154  

t$type <- "asymp"
t
#> 
#> 	 Multiple Comparison Based on Studentized Statistic 
#> 
#> scoring: none    type: asymp    method: bonferroni
#> 
#> family-wise confidence level: 95%
#> 
#>                           statistic     p-value  
#> location_1 ~ location_2  0.02737145 0.978434752  
#> location_1 ~ location_3 -1.77914406 0.090419680  
#> location_1 ~ location_4 -3.41048230 0.002773484 *
#> location_2 ~ location_3 -1.80651550 0.085909110  
#> location_2 ~ location_4 -3.43785375 0.002603664 *
#> location_3 ~ location_4 -1.63133824 0.118466249  

t$scoring <- "rank"
t
#> 
#> 	 Multiple Comparison Based on Studentized Statistic 
#> 
#> scoring: rank    type: asymp    method: bonferroni
#> 
#> family-wise confidence level: 95%
#> 
#>                          statistic     p-value  
#> location_1 ~ location_2  0.1307784 0.895950645  
#> location_1 ~ location_3 -1.4821547 0.138299135  
#> location_1 ~ location_4 -2.8335311 0.004603683 *
#> location_2 ~ location_3 -1.6129331 0.106759051  
#> location_2 ~ location_4 -2.9643094 0.003033631 *
#> location_3 ~ location_4 -1.3513764 0.176574906  

t$method <- "tukey"
t
#> 
#> 	 Multiple Comparison Based on Studentized Statistic 
#> 
#> scoring: rank    type: asymp    method: tukey
#> 
#> family-wise confidence level: 95%
#> 
#>                         statistic    p-value  
#> location_1 ~ location_2 0.1849485 0.99920177  
#> location_1 ~ location_3 2.0960833 0.44830284  
#> location_1 ~ location_4 4.0072180 0.02382224 *
#> location_2 ~ location_3 2.2810318 0.37122411  
#> location_2 ~ location_4 4.1921666 0.01604503 *
#> location_3 ~ location_4 1.9111348 0.53008595  

t$scoring <- "none"
t
#> 
#> 	 Multiple Comparison Based on Studentized Statistic 
#> 
#> scoring: none    type: asymp    method: tukey
#> 
#> family-wise confidence level: 95%
#> 
#>                          statistic    p-value  
#> location_1 ~ location_2 0.03870907 0.99999236  
#> location_1 ~ location_3 2.51608965 0.31194094  
#> location_1 ~ location_4 4.82315032 0.01356971 *
#> location_2 ~ location_3 2.55479873 0.29951264  
#> location_2 ~ location_4 4.86185939 0.01277452 *
#> location_3 ~ location_4 2.30706067 0.38453323  

t$type <- "permu"
t
#> 
#> 	 Multiple Comparison Based on Studentized Statistic 
#> 
#> scoring: none    type: permu(10000)    method: tukey
#> 
#> family-wise confidence level: 95%
#> 
#>                          statistic p-value  
#> location_1 ~ location_2 0.03870907  1.0000  
#> location_1 ~ location_3 2.51608965  0.3172  
#> location_1 ~ location_4 4.82315032  0.0150 *
#> location_2 ~ location_3 2.55479873  0.3052  
#> location_2 ~ location_4 4.86185939  0.0143 *
#> location_3 ~ location_4 2.30706067  0.3921