Ounces Of Beverage

The amounts of liquid in randomly selected beverage containers before and after the filling process has been repaired.

Usage

Table2.8.1

Format

An object of class data.frame with 5 rows and 2 columns.

Source

Table 2.8.1