Syntactic Sugar for Object Construction

Construct test objects in a unified way.

pmt(key, ...)

pmts(
  which = c("all", "onesample", "twosample", "ksample", "multicomp", "paired", "rcbd",
    "association", "table")
)

define_pmt(
  inherit = c("twosample", "ksample", "multicomp", "paired", "rcbd", "association",
    "table"),
  statistic = NULL,
  rejection = c("lr", "l", "r"),
  scoring = c("none", "rank", "vw", "expon"),
  n_permu = 0L,
  name = "User-Defined Permutation Test",
  alternative = NULL
)

Arguments

key

a character string corresponding to the desired test. Check pmts() for valid keys.

...

extra parameters passed to the constructor.

which

a character string specifying the desired tests.

inherit

a character string specifying the desired permutation test.

statistic

a function defining the test statistic.

rejection

a character string specifying where the rejection region is.

scoring, n_permu

passed to the constructor.

name

a character string specifying the name of the test.

alternative

a character string specifying the alternative of the test.