Syntactic Sugar for Object Construction

Construct test objects in a unified way.

Usage

pmt(key, ...)

pmts(
  which = c("all", "onesample", "twosample", "ksample", "multcomp", "paired", "rcbd",
    "association", "table")
)

define_pmt(
  statistic,
  inherit = c("twosample", "ksample", "paired", "rcbd", "association", "table"),
  rejection = c("lr", "l", "r"),
  scoring = c("none", "rank", "vw", "expon"),
  n_permu = 10000,
  name = "User-Defined Permutation Test",
  alternative = NULL,
  depends = character(),
  plugins = character(),
  includes = character()
)

Arguments

key

a character string specifying the test. Check pmts() for valid keys.

...

extra parameters passed to the constructor.

which

a character string specifying the desired tests.

statistic

definition of the test statistic. See Details.

inherit

a character string specifying the desired permutation test.

rejection

a character string specifying where the rejection region is.

scoring, n_permu

passed to the constructor.

name

a character string specifying the name of the test.

alternative

a character string specifying the alternative of the test.

depends, plugins, includes

passed to Rcpp::cppFunction().

Value

a test object corresponding to the specified key.

a data frame containing keys and corresponding tests implemented in this package.

Details

The test statistic in define_pmt can be defined using either R or Rcpp (C++14 required), with the statistic parameter specified as:

R: a function returning a closure that returns a double.

Rcpp: a character string defining a captureless lambda (introduced in C++11) returning another lambda that may capture by value, accepts const arguments of the same type, and returns a double.

When using Rcpp, the parameters for different inherit are listed as follows. Note that the parameter names are illustrative and may be modified.

"twosample": (Rcpp::NumericVector sample_1, Rcpp::NumericVector sample_2)

"ksample": (Rcpp::NumericVector combined_sample, Rcpp::IntegerVector one_based_group_index)

"paired": (Rcpp::NumericVector sample_1, Rcpp::NumericVector sample_2)

"rcbd": (Rcpp::NumericMatrix block_as_column_data)

"association": (Rcpp::NumericVector sample_1, Rcpp::NumericVector sample_2)

"table": (Rcpp::IntegerMatrix contingency_table)

Defining the test statistic using R follows a similar approach. The purpose of this design is to pre-calculate certain constants that remain invariant during permutation.

Examples

pmt("twosample.wilcoxon")
#> <Wilcoxon>
#>   Inherits from: <TwoSampleLocationTest>
#>   Public:
#>     alternative: active binding
#>     conf_int: active binding
#>     conf_level: active binding
#>     correct: active binding
#>     data: active binding
#>     estimate: active binding
#>     initialize: function (type = c("permu", "asymp"), alternative = c("two_sided", 
#>     method: active binding
#>     n_permu: active binding
#>     null_value: active binding
#>     p_value: active binding
#>     plot: function (style = c("graphics", "ggplot2"), ...) 
#>     print: function () 
#>     scoring: active binding
#>     statistic: active binding
#>     test: function (...) 
#>     type: active binding
#>   Private:
#>     .alternative: two_sided
#>     .autoplot: function (...) 
#>     .calculate: function () 
#>     .calculate_extra: function () 
#>     .calculate_n_permu: function () 
#>     .calculate_p: function () 
#>     .calculate_p_permu: function () 
#>     .calculate_score: function () 
#>     .calculate_side: function () 
#>     .calculate_statistic: function () 
#>     .compile_statistic_closure: function () 
#>     .conf_int: NULL
#>     .conf_level: 0.95
#>     .correct: TRUE
#>     .data: NULL
#>     .data_name: NULL
#>     .define: function () 
#>     .estimate: NULL
#>     .link: +
#>     .method: default
#>     .n_permu: 10000
#>     .name: Two-Sample Wilcoxon Test
#>     .null_value: 0
#>     .on_alternative_change: function () 
#>     .on_conf_level_change: function () 
#>     .on_method_change: function () 
#>     .on_n_permu_change: function () 
#>     .on_null_value_change: function () 
#>     .on_scoring_change: function () 
#>     .on_type_change: function () 
#>     .p_value: NULL
#>     .param_name: location shift
#>     .plot: function (...) 
#>     .preprocess: function () 
#>     .print: function () 
#>     .raw_data: NULL
#>     .scoring: rank
#>     .side: NULL
#>     .statistic: NULL
#>     .statistic_func: NULL
#>     .type: permu

pmts()
#>                     key              class
#> 1    onesample.quantile           Quantile
#> 2         onesample.cdf                CDF
#> 3  twosample.difference         Difference
#> 4    twosample.wilcoxon           Wilcoxon
#> 5    twosample.scoresum           ScoreSum
#> 6      twosample.ansari      AnsariBradley
#> 7      twosample.siegel        SiegelTukey
#> 8         twosample.rmd  RatioMeanDeviance
#> 9          twosample.ks  KolmogorovSmirnov
#> 10       ksample.oneway             OneWay
#> 11           ksample.kw      KruskalWallis
#> 12           ksample.jt JonckheereTerpstra
#> 13 multcomp.studentized        Studentized
#> 14          paired.sign               Sign
#> 15    paired.difference   PairedDifference
#> 16          rcbd.oneway         RCBDOneWay
#> 17        rcbd.friedman           Friedman
#> 18            rcbd.page               Page
#> 19     association.corr        Correlation
#> 20          table.chisq          ChiSquare
#>                                                  test
#> 1                                       Quantile Test
#> 2       Inference on Cumulative Distribution Function
#> 3             Two-Sample Test Based on Mean or Median
#> 4                            Two-Sample Wilcoxon Test
#> 5              Two-Sample Test Based on Sum of Scores
#> 6                                 Ansari-Bradley Test
#> 7                                   Siegel-Tukey Test
#> 8                            Ratio Mean Deviance Test
#> 9                  Two-Sample Kolmogorov-Smirnov Test
#> 10                       One-Way Test for Equal Means
#> 11                                Kruskal-Wallis Test
#> 12                           Jonckheere-Terpstra Test
#> 13 Multiple Comparison Based on Studentized Statistic
#> 14                               Two-Sample Sign Test
#> 15             Paired Comparison Based on Differences
#> 16               One-Way Test for Equal Means in RCBD
#> 17                                      Friedman Test
#> 18                                          Page Test
#> 19        Test for Association Between Paired Samples
#> 20               Chi-Square Test on Contingency Table

r <- define_pmt(
    inherit = "twosample", rejection = "lr", n_permu = 1e5,
    statistic = function(x, y) {
        m <- length(x)
        n <- length(y)
        function(x, y) sum(x) / m - sum(y) / n
    }
)

cpp <- define_pmt(
    inherit = "twosample", rejection = "lr", n_permu = 1e5,
    statistic = "[](NumericVector x, NumericVector y) {
        R_len_t n_x = x.size();
        R_len_t n_y = y.size();
        return [n_x, n_y](const NumericVector x, const NumericVector y) {
            double sum_x = 0;
            double sum_y = 0;
            for(auto x_i : x) sum_x += x_i;
            for(auto y_i : y) sum_y += y_i;
            return sum_x / n_x - sum_y / n_y;
        };
    }"
)

x <- rnorm(100)
y <- rnorm(100, 1)
options(LearnNonparam.pmt_progress = FALSE)
system.time(r$test(x, y))
#>    user  system elapsed 
#>   0.328   0.000   0.328 
system.time(cpp$test(x, y))
#>    user  system elapsed 
#>    0.21    0.00    0.21